pyproject.toml
[project]
name = "phara"
version = "0.1.0"
description = "A chatbot that integrates with Gemini AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "requests>=2.28.0",
    "pandas>=2.0.0",
    "openpyxl>=3.0.0"
]
